add confirmation before user logout ref - self

+++ js/common.js
@@ -32,6 +32,15 @@ function setActiveMenu() {
 }
 
 function logout() {
+
+    const confirmLogout = confirm(
+        "Are you sure you want to logout?"
+    );
+
+    if (!confirmLogout) {
+        return;
+    }
+
     localStorage.clear();
     sessionStorage.clear();
